Remove usePDFExport hook
